Check if queue exists and update attributes if so
If the queue already exists, Amazon SQS will throw a http-400 unless both name and all attributes match from the request with the existing queue.
Thus if we change the default _peekLockDuration we will receive an error when trying to create the queue.

Instead we first check for the existence of a queue with the same name, and simply update the VisibilityTimeout if so, otherwise we create a new queue with the desired default VisibilityTimeout.